Bump version without building
Updated release-recurring and release-core-build to use bump version, and remove old release script.
