+++ .github/workflows/release-recurring.yml
@@ -39,9 +39,13 @@ jobs:
       - name: Bootstrap packages
         run: npm run bootstrap
 
+      - name: Bump version
+        shell: bash
+        run: BUILD_REF="$(git rev-parse --short HEAD)${{ github.event_name == 'pull_request' && '.pr-$PR_NUMBER' || '' }}" npm run app-bump-version
+
       - name: Package
         shell: bash
-        run: BUILD_REF="$(git rev-parse --short HEAD)${{ github.event_name == 'pull_request' && '.pr-$PR_NUMBER' || '' }}" BUILD_TARGETS='${{ matrix.build-targets }}' npm run app-package
+        run: BUILD_TARGETS='${{ matrix.build-targets }}' npm run app-package
 
       - name: Upload artifacts
         uses: actions/upload-artifact@v2
