+++ package.json
@@ -18,9 +18,6 @@
     "bootstrap": "npm install && lerna bootstrap && lerna run --stream bootstrap",
     "bootstrap:electron-builder": "lerna bootstrap --no-ci -- --production --no-optional",
     "version": "lerna version --exact",
-    "version:deprecated": "lerna version --exact --preid alpha",
-    "version:all": "npm run version:deprecated -- --force-publish",
-    "version:dry": "npm run version:deprecated -- --no-git-tag-version",
     "release": "lerna publish from-git --pre-dist-tag beta --no-verify-access --yes",
     "clean": "lerna run clean --parallel --stream && lerna clean --yes && rimraf node_modules",
     "test": "lerna run --stream --parallel test",
@@ -31,7 +28,7 @@
     "app-start": "npm start --prefix packages/insomnia-app",
     "app-build": "npm run build --prefix packages/insomnia-app",
     "app-package": "npm run package --prefix packages/insomnia-app",
-    "app-release": "npm run release --prefix packages/insomnia-app",
+    "app-bump-version": "npm run bump-version --prefix packages/insomnia-app",
     "preinstall": "node ./scripts/check-version.js",
     "app-build:smoke": "cross-env npm run app-build",
     "app-package:smoke": "cross-env npm run app-package",
